Completed state.dat tests
Successfully wrote state.dat with appropriate values;
Verified that state.dat is ignored & overwritten (most likely by the BS2
state machine, which is part of SYSMENU);
Added in some function to dump some memory to SD, to see if I can
pinpoint where the value indicating that the console was booted from the
"Eject" button, as opposed to the regular "Power" button, is stored. (No
luck so far).